Refactor completion
